Add files via upload
We get an error, 
with message Error: Response is constant across permutations. Please check your model. One of many reason it throws an error is because the input for our model is a document-term matrix, with many sparse/empty values. LIME don’t work really with an input that contain many sparse value like in text classification3. With the text contain only the word yes, the matrix will be 0 (Absent) or sparse for most of the feature and 1 (Present) for a single feature, that is the term yes. So, whenever you found an error like this, you can be certain that the input is a matrix with empty values. To my best knowledge, there is no workaround for this problem yet. Therefore, we only concern with a review or text that contain several words.